Allow freetype to be unbundled.
